# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Climbr API",
    "description": "API for Climbr - Career Platform connecting young African talent with job opportunities and training programs",
    "version": "1.0.0"
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Jobs",
        "description": "Get all active job listings with advanced filtering and pagination",
        "operationId": "get_jobs_jobs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in title, description, or company",
              "title": "Search"
            },
            "description": "Search in title, description, or company"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by location",
              "title": "Location"
            },
            "description": "Filter by location"
          },
          {
            "name": "job_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by job type",
              "title": "Job Type"
            },
            "description": "Filter by job type"
          },
          {
            "name": "industry",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by industry",
              "title": "Industry"
            },
            "description": "Filter by industry"
          },
          {
            "name": "experience_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by experience level",
              "title": "Experience Level"
            },
            "description": "Filter by experience level"
          },
          {
            "name": "company_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by company size",
              "title": "Company Size"
            },
            "description": "Filter by company size"
          },
          {
            "name": "salary_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum salary",
              "title": "Salary Min"
            },
            "description": "Minimum salary"
          },
          {
            "name": "salary_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum salary",
              "title": "Salary Max"
            },
            "description": "Maximum salary"
          },
          {
            "name": "skills",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by skill IDs",
              "title": "Skills"
            },
            "description": "Filter by skill IDs"
          },
          {
            "name": "date_posted",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by date posted (recent, week, month)",
              "title": "Date Posted"
            },
            "description": "Filter by date posted (recent, week, month)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort by: date, salary, relevance",
              "default": "date",
              "title": "Sort By"
            },
            "description": "Sort by: date, salary, relevance"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Jobs Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Job Details",
        "description": "Get details for a specific job",
        "operationId": "get_job_details_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/employer/purchase": {
      "post": {
        "tags": [
          "employer"
        ],
        "summary": "Purchase Credits",
        "description": "Initialize a Paystack transaction to purchase job posting credits",
        "operationId": "purchase_credits_employer_purchase_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_purchase_credits_employer_purchase_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/employer/confirm-payment": {
      "post": {
        "tags": [
          "employer"
        ],
        "summary": "Confirm Payment",
        "description": "Verify a Paystack transaction and credit the employer account",
        "operationId": "confirm_payment_employer_confirm_payment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_confirm_payment_employer_confirm_payment_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/employer/credits": {
      "get": {
        "tags": [
          "employer"
        ],
        "summary": "Get Credits",
        "description": "Get current job posting credits",
        "operationId": "get_credits_employer_credits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/employer/jobs": {
      "post": {
        "tags": [
          "employer"
        ],
        "summary": "Create Job",
        "description": "Create a new job posting (requires credits)",
        "operationId": "create_job_employer_jobs_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "employer"
        ],
        "summary": "Get Employer Jobs",
        "description": "Get all jobs posted by the current employer",
        "operationId": "get_employer_jobs_employer_jobs_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/employer/jobs/{job_id}/image": {
      "post": {
        "tags": [
          "employer"
        ],
        "summary": "Upload Job Image",
        "description": "Upload an image for a job posting",
        "operationId": "upload_job_image_employer_jobs__job_id__image_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_job_image_employer_jobs__job_id__image_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_confirm_payment_employer_confirm_payment_post": {
        "properties": {
          "reference": {
            "type": "string",
            "title": "Reference"
          }
        },
        "type": "object",
        "required": [
          "reference"
        ],
        "title": "Body_confirm_payment_employer_confirm_payment_post"
      },
...
      "Body_purchase_credits_employer_purchase_post": {
        "properties": {
          "package_id": {
            "type": "integer",
            "title": "Package Id"
          }
        },
        "type": "object",
        "required": [
          "package_id"
        ],
        "title": "Body_purchase_credits_employer_purchase_post"
      },
...
      "Body_upload_job_image_employer_jobs__job_id__image_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_job_image_employer_jobs__job_id__image_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}